+++ packages/web-console/src/store/Query/types.ts
@@ -34,6 +34,8 @@ export enum NotificationType {
   ERROR = "error",
   INFO = "info",
   SUCCESS = "success",
+  NOTICE = "notice",
+  LOADING = "loading",
 }
 
 export type NotificationShape = Readonly<{