+++ packages/web-console/src/utils/questdb.ts
@@ -27,6 +27,7 @@ import { eventBus } from "../modules/EventBus"
 import { EventType } from "../modules/EventBus/types"
 import { AuthPayload } from "../modules/OAuth2/types"
 import { StoreKey } from "./localStorage/types"
+import {API_VERSION} from "../consts";
 
 type ColumnDefinition = Readonly<{ name: string; type: string }>
 
@@ -38,6 +39,7 @@ export enum Type {
   DML = "dml",
   DQL = "dql",
   ERROR = "error",
+  NOTICE = "notice",
 }
 
 export type Timings = {
@@ -58,6 +60,7 @@ type RawDqlResult = {
   dataset: DatasetType[]
   ddl: undefined
   dml: undefined
+  notice: undefined
   error: undefined
   query: string
   timings: Timings
@@ -82,6 +85,15 @@ type RawErrorResult = {
   query: string
 }
 
+type RawNoticeResult = {
+    ddl: undefined
+    dml: undefined
+    error: undefined
+    notice: "<notice message>"
+    position: undefined
+    query: string
+}
+
 type DdlResult = {
   query: string
   type: Type.DDL
@@ -92,11 +104,15 @@ type DmlResult = {
   type: Type.DML
 }
 
-type RawResult = RawDqlResult | RawDmlResult | RawDdlResult | RawErrorResult
+type RawResult = RawDqlResult | RawDmlResult | RawDdlResult | RawErrorResult | RawNoticeResult
 
 export type ErrorResult = RawErrorResult & {
   type: Type.ERROR
   status: number
+}
+
+export type NoticeResult = RawNoticeResult & {
+    type: Type.NOTICE
 }
 
 export type QueryRawResult =
@@ -104,6 +120,7 @@ export type QueryRawResult =
   | DmlResult
   | DdlResult
   | ErrorResult
+  | NoticeResult
 
 export type QueryResult<T extends Record<string, any>> =
   | {
@@ -117,6 +134,7 @@ export type QueryResult<T extends Record<string, any>> =
   | ErrorResult
   | DmlResult
   | DdlResult
+  | NoticeResult
 
 export type PartitionBy = "HOUR" | "DAY" | "WEEK" | "MONTH" | "YEAR" | "NONE"
 
@@ -401,6 +419,7 @@ export class Client {
       src: "con",
       query,
       timings: true,
+      version: API_VERSION,
       ...options,
     }
 
@@ -503,6 +522,13 @@ export class Client {
         })
       }
 
+      if (data.notice) {
+        return {
+          ...data,
+          type: Type.NOTICE,
+        }
+      }
+
       return {
         ...data,
         timings: {
@@ -594,6 +620,7 @@ export class Client {
         `chk?${Client.encodeParams({
           f: "json",
           j: name,
+          version: API_VERSION,  
         })}`,
         { headers: this.commonHeaders },
       )
@@ -667,7 +694,7 @@ export class Client {
   async exportQueryToCsv(query: string) {
     try {
       const response: Response = await fetch(
-        `exp?${Client.encodeParams({ query })}`,
+        `exp?${Client.encodeParams({ query, version: API_VERSION })}`,
         { headers: this.commonHeaders },
       )
       const blob = await response.blob()